# gozero/application/gateway/internal/openapi/swagger.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "info": {
    "description": "go-zero implementation of the native HTTP contract",
    "title": "microservice-kit API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/operLog": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "operationId": "operlogOperLogCreate",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "businessType": {
                  "type": "string"
                },
                "costTime": {
                  "type": "integer"
                },
                "deviceType": {
                  "type": "string"
                },
                "errorMsg": {
                  "type": "string"
                },
                "jsonResult": {
                  "type": "string"
                },
                "method": {
                  "type": "string"
                },
                "operIp": {
                  "type": "string"
                },
                "operLocation": {
                  "type": "string"
                },
                "operName": {
                  "type": "string"
                },
                "operParam": {
                  "type": "string"
                },
                "operUrl": {
                  "type": "string"
                },
                "requestMethod": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "userAgent": {
                  "type": "string"
                }
              },
              "required": [
                "title"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "OperLogCreate"
      }
    },
...
    "/api/v1/operLog/clean": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "operationId": "operlogOperLogClean",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "days": {
                  "type": "integer"
                }
              },
              "required": [
                "days"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "OperLogClean"
      }
    },
    "/api/v1/operLog/page": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "operationId": "operlogOperLogPage",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "businessType": {
                  "type": "string"
                },
                "endTime": {
                  "type": "string"
                },
                "operName": {
                  "type": "string"
                },
                "pageNum": {
                  "type": "integer"
                },
                "pageSize": {
                  "type": "integer"
                },
                "startTime": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "OperLogPage"
      }
    },
...
    "/auth/refresh": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "operationId": "authRefreshToken",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "clientId": {
                  "type": "string"
                },
                "refreshToken": {
                  "type": "string"
                }
              },
              "required": [
                "refreshToken",
                "clientId"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "RefreshToken"
      }
    },
...
    "/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "operationId": "authLogin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "appid": {
                  "type": "string"
                },
                "clientId": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "grantType": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "phonenumber": {
                  "type": "string"
                },
                "smsCode": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "uuid": {
                  "type": "string"
                },
                "wxCode": {
                  "type": "string"
                }
              },
              "required": [
                "clientId",
                "grantType"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "Login"
      }
    },
...
    "/resource/sms/code": {
      "get": {
        "operationId": "captchaResourceSmsCode",
        "parameters": [
          {
            "allowEmptyValue": true,
            "in": "query",
            "name": "phone",
            "type": "string"
          },
          {
            "allowEmptyValue": true,
            "in": "query",
            "name": "phonenumber",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "code": {
                  "type": "integer"
                },
                "data": {},
                "msg": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "schemes": [
          "https"
        ],
        "summary": "ResourceSmsCode"
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}